SmartDialer: show quick contact on long press
Instead of displaying the disambiguation dialog, show quick contact
instead which also allows the user to decide which number to call as
well as perform other operations such as editing or messaging.

This is useful because it's easier to search for contacts via T9
compared to using the QWERTY keyboard.

Change-Id: I81c21e58fb1090157a062d79600287d7e1e7c8d6